moving topojson layer to file, adding in tile stuff for PR
