+++ api/models/Tiles.js
@@ -1,8 +1,8 @@
 var Tilenik = require('tilenik'),
-  //nodetiles = require('nodetiles-core'),
-  //GeoJsonSource = nodetiles.datasources.GeoJson,
-  //Projector = nodetiles.projector,
-  //Canvas = require('canvas'),
+  nodetiles = require('nodetiles-core'),
+  GeoJsonSource = nodetiles.datasources.GeoJson,
+  Projector = nodetiles.projector,
+  Canvas = require('canvas'),
   nfs = require('node-fs'),
   fs = require('fs'),
   terraformer = require('Terraformer'),
