+++ views/layout.ejs
@@ -21,6 +21,7 @@
     <script type="text/javascript" src="/js/bootstrap.min.js"></script>
     <script type="text/javascript" src="/js/esri-leaflet.min.js"></script>
     <script type="text/javascript" src="/js/topojson.v1.min.js"></script>
+    <script type="text/javascript" src="/js/L.TopoJSON.js"></script>
     <script type="text/javascript" src="/linker/js/map.js"></script>
 	
     <%- body %>
